5.28

+++ USER/inc/DJI_test.h
@@ -5,7 +5,7 @@
 void USART2_3_NVIC_Configuration(void);
 void usart2_3_init(void);
 void USART_Send_single(uint8_t usart,uint8_t date);
-void USART_Send_number_5(unsigned int number);
+void USART_Send_number_5(int16_t number);
 void USART_Send_number_3(int16_t number);
 void USART2_Send_data(uint8_t* str,uint8_t len);
 #endif
